Refactor Node/Platform CVE tables
